+++ lib/injected/build.gradle
@@ -1,4 +1,3 @@
-
 dependencies {
     // All these dependencies will be provided at runtime by the service this
     // library is injected into.
@@ -12,10 +11,10 @@ dependencies {
     compileOnly 'org.springframework.security:spring-security-oauth2-jose:5.4.9'
 
 
-    compileOnly ('com.github.hmcts:idam-java-client:2.0.1') {
+    compileOnly('com.github.hmcts:idam-java-client:2.0.1') {
         transitive = false
     }
-    compileOnly ('com.github.hmcts:auth-checker-lib:2.1.4') {
+    compileOnly('com.github.hmcts:auth-checker-lib:2.1.4') {
         transitive = false
     }
     // This is needed at runtime to decode JWTs
