+++ lib/injected/src/main/java/uk/gov/hmcts/rse/ccd/lib/FailFast.java
@@ -6,12 +6,12 @@ import org.springframework.boot.diagnostics.FailureAnalysis;
 // Fast shutdown the JVM (ie. all services) if any
 // individual service fails to boot.
 public class FailFast extends AbstractFailureAnalyzer {
-  @Override
-  protected FailureAnalysis analyze(Throwable rootFailure, Throwable cause) {
-    cause.printStackTrace();
-    System.out.println("**** CFTLIB failed to start ****");
-    Runtime.getRuntime().halt(-1);
-    // Unreachable
-    throw new RuntimeException();
-  }
+    @Override
+    protected FailureAnalysis analyze(Throwable rootFailure, Throwable cause) {
+        cause.printStackTrace();
+        System.out.println("**** CFTLIB failed to start ****");
+        Runtime.getRuntime().halt(-1);
+        // Unreachable
+        throw new RuntimeException();
+    }
 }
