+++ lib/injected/src/main/java/uk/gov/hmcts/rse/ccd/lib/S2SLib.java
@@ -1,6 +1,5 @@
 package uk.gov.hmcts.rse.ccd.lib;
 
-
 import java.util.Date;
 import java.util.Map;
 
@@ -13,23 +12,8 @@ import org.aspectj.lang.annotation.Aspect;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
-import org.springframework.context.annotation.Configuration;
 import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.auth.parser.idam.core.service.token.ServiceTokenParser;
 import uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi;
-
-@Configuration
-@ConditionalOnClass(ServiceTokenParser.class)
-@Aspect
-class NonValidatingS2SParser {
-
-    // Intercept S2S validation requests and parse the token locally without validating the signature.
-    // This allows self-signed s2s tokens and breaks the dependency on s2s.
-    @Around("execution(* uk.gov.hmcts.reform.auth.parser.idam.core.service.token.ServiceTokenParser.parse(..)) && args(jwt)")
-    public Object parse(ProceedingJoinPoint p, String jwt) throws Throwable {
-        return JWT.decode(jwt.replace("Bearer ", "")).getSubject();
-    }
-}
 
 @Component
 @ConditionalOnClass(ServiceAuthorisationApi.class)
@@ -42,13 +26,14 @@ class S2SLib {
     @Autowired
     public S2SLib(@Value("${idam.s2s-auth.microservice}") String service,
                   @Value("${rse.lib.stub.auth.outbound:false}") boolean stubOutbound
-                  ) {
+    ) {
         this.service = service;
         this.stubOutbound = stubOutbound;
     }
 
     @SneakyThrows
-    @Around("execution(* uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi.getServiceName(..)) && args(authHeader)")
+    @Around("execution(* uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi.getServiceName(..))"
+        + " && args(authHeader)")
     public Object getServiceName(ProceedingJoinPoint p, String authHeader) {
         if (null != authHeader) {
             var subject = JWT.decode(authHeader.replace("Bearer ", "")).getSubject();
@@ -59,20 +44,20 @@ class S2SLib {
             }
             return subject;
         }
-       return p.proceed();
+        return p.proceed();
     }
 
     @SneakyThrows
     @Around("execution(* uk.gov.hmcts.reform.authorisation.ServiceAuthorisationApi.serviceToken(..)) && args(signIn)")
     public Object serviceToken(ProceedingJoinPoint p, Map<String, String> signIn) {
         if (stubOutbound) {
-          // Return a self-signed JWT.
-          return JWT.create()
-            .withSubject(signIn.get("microservice"))
-            .withIssuedAt(new Date())
-            .sign(Algorithm.HMAC256("secret"));
+            // Return a self-signed JWT.
+            return JWT.create()
+                .withSubject(signIn.get("microservice"))
+                .withIssuedAt(new Date())
+                .sign(Algorithm.HMAC256("secret"));
         }
 
-      return p.proceed();
+        return p.proceed();
     }
 }
