+++ lib/injected/src/main/java/uk/gov/hmcts/rse/ccd/lib/URLRewriter.java
@@ -1,13 +1,12 @@
 package uk.gov.hmcts.rse.ccd.lib;
 
-import java.io.IOException;
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletRequestWrapper;
 import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
 
-import com.auth0.jwt.JWT;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
@@ -20,7 +19,7 @@ import uk.gov.hmcts.reform.idam.client.IdamApi;
 /**
  * Implements URL remappings handled by the CCD API Gateway.
  */
-@Order(value= Ordered.HIGHEST_PRECEDENCE)
+@Order(value = Ordered.HIGHEST_PRECEDENCE)
 @ConditionalOnClass(IdamApi.class)
 @Component
 public class URLRewriter extends OncePerRequestFilter {
@@ -32,7 +31,7 @@ public class URLRewriter extends OncePerRequestFilter {
 
     @Autowired
     public URLRewriter(IdamApi idam,
-      @Value("${rse.lib.service_name:***CFT lib***}") String name) {
+                       @Value("${rse.lib.service_name:***CFT lib***}") String name) {
         this.idam = idam;
         this.name = name;
     }
@@ -40,13 +39,13 @@ public class URLRewriter extends OncePerRequestFilter {
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
         throws ServletException, IOException {
-      String name = Thread.currentThread().getName();
-      try {
-        Thread.currentThread().setName("*** " + this.name);
-        filterChain.doFilter(new Rewriter(request), response);
-      } finally {
-        Thread.currentThread().setName(name);
-      }
+        String name = Thread.currentThread().getName();
+        try {
+            Thread.currentThread().setName("*** " + this.name);
+            filterChain.doFilter(new Rewriter(request), response);
+        } finally {
+            Thread.currentThread().setName(name);
+        }
     }
 
     class Rewriter extends HttpServletRequestWrapper {
@@ -67,7 +66,7 @@ public class URLRewriter extends OncePerRequestFilter {
         private String process(String url) {
             // CCD Gateway strips this path.
             if (url.startsWith("/data")) {
-              url = url.replaceFirst("/data", "");
+                url = url.replaceFirst("/data", "");
             }
             // Gateway replaces placeholder userIDs.
             if (url.contains("/:uid/")) {
