+++ lib/rse-cft-lib/src/main/java/uk/gov/hmcts/rse/ccd/lib/Project.java
@@ -1,9 +1,9 @@
 package uk.gov.hmcts.rse.ccd.lib;
 
 public enum Project {
-  Datastore,
-  Definitionstore,
-  Userprofile,
-  AM,
-  Application
+    Datastore,
+    Definitionstore,
+    Userprofile,
+    AM,
+    Application
 }
